+++ src/Models/Material/Material_Models/Correspondence/Correspondence.jl
@@ -59,7 +59,6 @@ function init_model(
     datamanager.create_constant_node_field("Strain Increment", Float64, "Matrix", dof)
     datamanager.create_node_field("Cauchy Stress", Float64, "Matrix", dof)
     datamanager.create_node_field("von Mises Stress", Float64, 1)
-    datamanager.create_constant_node_field("Angles", Float64, dof)
     rotation::Bool = datamanager.get_rotation()
     material_models = split(material_parameter["Material Model"], "+")
     material_models = map(r -> strip(r), material_models)